+++ src/server/element.cpp
@@ -3,7 +3,7 @@
 namespace websocket {
 	int Element::idCount_;
 
-	Element::Element(int & x,int & y,int & radius): id_(++idCount_), x_(x), y_(y),radius_(radius)
+	Element::Element(double & x,double & y,double & radius): id_(++idCount_), x_(x), y_(y),radius_(radius)
 	{
 		
 	}
