+++ src/server/foodItem.cpp
@@ -1,7 +1,7 @@
 #include "foodItem.hpp"
 
 namespace websocket{
-    FoodItem::FoodItem(int & x, int & y,int radius): Element(x,y,radius)
+    FoodItem::FoodItem(double & x, double & y,double radius): Element(x,y,radius)
     {
         //std::cout << "construct food" << std::endl;
     }
